+++ src/guards/string.guard.ts
@@ -124,6 +124,7 @@ export class StringGuard extends Guard<StringRule> {
      * {@link https://en.wikipedia.org/wiki/List_of_precomposed_Latin_characters_in_Unicode | 'precomposed-latin'} |
      * {@link https://en.wikipedia.org/wiki/Danish_orthography | 'dan'} |
      * {@link https://en.wikipedia.org/wiki/German_orthography | 'deu'} |
+     * {@link https://en.wikipedia.org/wiki/Estonian_orthography | 'est'} |
      * {@link https://en.wikipedia.org/wiki/Finnish_orthography | 'fin'} |
      * {@link https://en.wikipedia.org/wiki/French_orthography | 'fra'} |
      * {@link https://en.wikipedia.org/wiki/Irish_orthography | 'gle'} |