+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.5.0] - 2021-06-09
+## [2.5.0] - 2021-06-15
 
 ### Added
 
--   Add optional parameters 'dan', 'fin', 'gle', nld', and 'pol' to Tyr.string().isAlpha() method.
+-   Add optional parameters 'dan', 'est', 'fin', 'gle', nld' and 'pol' to Tyr.string().isAlpha() method.
 
 ## [2.4.0] - 2021-06-03